wip: rendering, validation and styling logic for form.
Have manually tested and checked that focus, active and error
states for the rendering of the form input fields styling is
working.

Logic to handle validation seems sound to me.

Need to add logic that clears and resets the form fields
and form data state if all form validation checks are passed.